refactored DailyArtPromptApp project structure and added logging in dap-service
