+++ src/__tests__/Comment/Comment.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Comment from './Comment'
+import Comment from '../../main/Comment/Comment'
 
 const defaultProps = {
   comment: 'text',
