+++ src/__tests__/Image/Image.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import { Image, mapDispatchToProps } from './Image'
-import * as TYPES from '../store/actions'
+import { Image, mapDispatchToProps } from '../../main/Image/Image'
+import * as TYPES from '../../main/store/actions'
 
 const defaultProps = {
   image: {
