+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
-import EntryScreen from './EntryScreen'
-import { store } from './store/rootReducer'
+import EntryScreen from './main/DailyArtPromptApp/EntryScreen'
+import { store } from './main/store/rootReducer'
 
 ReactDOM.render(
   <Provider store={store}>
