--- src/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  moduleNameMapper: {
-    '\\.css$': 'identity-obj-proxy',
-  },
-}
