+++ src/main/DailyArtPromptApp/main.css
