+++ src/main/Prompt/dateReducer.js
@@ -1,0 +1,13 @@
+import moment from 'moment'
+import * as TYPES from '../store/actions'
+
+export const dateReducer = (state = null, action) => {
+  switch (action.type) {
+    case TYPES.SET_INITIAL_DATE:
+      return moment().format('YYYY-MM-DD')
+    case TYPES.UPDATE_DATE:
+      return action.payload.date
+    default:
+      return state
+  }
+}
