--- src/webpack.config.js
@@ -1,4 +1,0 @@
-{
-  test: /\.css$/,
-  loader: "style-loader!css-loader?modules=true"
-}
